[Swift shipping] Add `Swift.Bindings.Experimental` nuget to `dotnet-experimental` for device tests (#3014)
This PR adds the Swift.Bindings.Experimental nuget package to the dotnet-experimental feed for device testing. The package provides both C# and Swift bindings for selected frameworks.

It will be used on Helix for testing StoreKit scenarios and other frameworks in the future.

The generate.sh script is updated to generate an .xcframework that contains platform and architecture specific bindings. The publish job is updated to upload Swift.Bindings.Experimental artifacts to the storage.